add concrete BleakScanner and BleakClient classes
This allows us to put all of the documentation, including platform-
specific quirks, for BleakScanner and BleakClient in one place. Docs
for these classes are move to new sub-pages of the API docs page and
the backend docs are moved to the backend pages of the docs.

It will also allow us to eventually eliminate quite a bit of duplicated
code from the backends by moving it to these shared classes.

Several methods that provided duplicate functionality have been
deprecated (previously they were required because of backend
implementation details but can now be safely removed from the
top-level classes).